Rename clp_main to run; Reorder some includes.
