+++ components/core/src/clp/clp.cpp
@@ -1,12 +1,13 @@
 #include <string>
+
 #include <spdlog/spdlog.h>
 
-#include "clp_main.hpp"
+#include "run.hpp"
 
 int main (int argc, const char* argv[]) {
     std::string archive_path;
     try {
-        return clp_main(argc, argv, &archive_path);
+        return clp::run(argc, argv, &archive_path);
     } catch (std::string const err) {
         SPDLOG_ERROR(err.c_str());
         return 1;
