--- components/core/src/clp/clp_main.hpp
@@ -1,9 +1,0 @@
-#ifndef CLP_MAIN_HPP
-#define CLP_MAIN_HPP
-
-// C++ libraries
-#include <string>
-
-int clp_main (int argc, const char* argv[], std::string* archive_path);
-
-#endif // CLP_MAIN_HPP
