+++ components/core/src/clp/run.hpp
@@ -1,0 +1,10 @@
+#ifndef CLP_RUN_HPP
+#define CLP_RUN_HPP
+
+#include <string>
+
+namespace clp {
+    int run (int argc, const char* argv[], std::string* archive_path);
+}
+
+#endif // CLP_RUN_HPP
